UTY-360: Add Support for Nested Type and Enum Generation (#45)
## NOTE: The last commit is all the generated test code. This is to allow easy diffing. Ignore this commit if you are not interested.

#### Description
A number of changes had to be made to enable recursively generating types:
1. The `enumSet` of all enums in a schema corpus now includes enums defined in types.
2. Refactored `UnityTypeGenerator` into two parts:
 a. `UnityTypeContent` which contains the template for the actual struct content. It recursively calls generate on itself and `UnityEnumContent`
 b. `UnityTypeGenerator` which contains the imports and namespace declaration and then generates the content using the above tt file.
3. `UnityEnumGenerator` had the same refactor as above.
4. Added an `IndentString` function to enable proper alignment of nested types.
5. Committed generated code to allow for easy diffing when we change the code generator.

Also uncommented the nested schema to show that it works! 
#### Tests
Ran `End2End` tests on the CodeGenerator. It compiled and ran properly!
#### Documentation
N/A - this is a change that gets us closer to full schema compatibility.
#### Primary reviewers
If your change will take a long time to review, you can name at most two primary reviewers who are ultimately responsible for reviewing this request. @ mention them.
@jessicafalk @samiwh