Actually Run Tests in CI (#382)
#### Description
TL;DR - Adding `Generate` as a `delayCall` on startup means that when CI runs, the GDK assemblies actually fail to build - but this doesn't cause Unity to exit with a non-zero exit code 😭. It will silently ignore the compilation failure and run tests from any successfully built assemblies.

This PR reverts the that change introduced in #286. This required moving the "GDK Tools Configuration" from a `ScriptableObject` to a raw JSON file - as the `AssetDatabase` is not available during compilation.

Swapped the custom inspector view for the `ScriptableObject` to an Editor window.
#### Tests
Ran `ci/test.sh` from a clean state and made sure that tests were actually picked up.
#### Documentation
N/A
#### Primary reviewers
If your change will take a long time to review, you can name at most two primary reviewers who are ultimately responsible for reviewing this request. @ mention them.
@jessicafalk @jared-improbable